Refactored: separated environment setup during verification
